refactor: adapt API routes to Next.js format for v0 compatibility
