+++ app/api/auth/logout/route.ts
@@ -1,0 +1,8 @@
+import { cookies } from 'next/headers';
+import { NextResponse } from 'next/server';
+
+export async function POST() {
+  const response = NextResponse.json({ message: 'Logged out successfully' });
+  response.cookies.delete('auth_session');
+  return response;
+}
