+++ app/api/auth/user/route.ts
@@ -1,0 +1,2 @@
+// Re-export from status route - same functionality
+export { GET } from '../status/route';
