--- vercel.json
@@ -1,18 +1,0 @@
-{
-  "$schema": "https://openapi.vercel.sh/vercel.json",
-  "version": 2,
-  "framework": "vite",
-  "buildCommand": "vite build",
-  "outputDirectory": "dist/public",
-  "installCommand": "pnpm install",
-  "rewrites": [
-    {
-      "source": "/api/(.*)",
-      "destination": "/api/[...path]"
-    },
-    {
-      "source": "/((?!api|assets|.*\\.ico|.*\\.png|.*\\.svg|.*\\.webmanifest|.*\\.txt|.*\\.xml|.*\\.json).*)",
-      "destination": "/index.html"
-    }
-  ]
-}
